Quality of Life Update
Just setting up things as they might actually remain for a bit to get an idea of what size the sprites would need to be.

+++ Scribble_Tactics/Assets/Scripts/CameraLogic.cs
@@ -9,10 +9,10 @@ public class Camera_Follow : MonoBehaviour
     [SerializeField] private CinemachineVirtualCamera cinemachineVirtualCamera;
 
     //Boundary variables for camera
-    [SerializeField] private int xBoundaryNegative = -25;
-    [SerializeField] private int yBoundaryNegative = -25;
-    [SerializeField] private int xBoundaryPositive = 25;
-    [SerializeField] private int yBoundaryPositive = 25;
+    [SerializeField] private float xBoundaryNegative = -25;
+    [SerializeField] private float yBoundaryNegative = -25;
+    [SerializeField] private float xBoundaryPositive = 25;
+    [SerializeField] private float yBoundaryPositive = 25;
 
     //Only for WASD movement speed
     [SerializeField] private float moveSpeed = 6f;
